Task/des 2905 cite this data type other (#1299)
* Add citation download options and metrics to ProjectPreview and PublishedDetailLayout

* Refactor Download Citation link in ProjectPreview and PublishedDetailLayout

* Refactor useCitationMetrics hook and remove unused error variable

* Refactor useCitationMetrics hook and update ProjectPreview module and styles

* Refactor useCitationMetrics hook and update ProjectPreview module and styles

* Refactor useCitationMetrics hook and remove MetricsModal from ProjectPreview module

* Refactor useCitationMetrics hook and remove unused error variable

* Refactor import path for MetricsModal in ProjectCitation and PublishedDetailLayout
