+++ client/modules/datafiles/src/projects/ProjectCitation/ProjectCitation.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   useCitationMetrics,
   useProjectDetail,
   usePublicationDetail,
 } from '@client/hooks';
-// import { MetricsModal } from '../modals/MetricsModal';
+import { MetricsModal } from '../modals/MetricsModal';
 import styles from './ProjectCitation.module.css';
 
 export const ProjectCitation: React.FC<{
@@ -91,6 +91,8 @@ export const DownloadCitation: React.FC<{
     error: projectError,
   } = usePublicationDetail(projectId);
 
+  const [isModalVisible, setIsModalVisible] = useState(false);
+
   const entityDetails = (data?.tree.children ?? []).find(
     (child) => child.uuid === entityUuid
   );
@@ -101,6 +103,14 @@ export const DownloadCitation: React.FC<{
       : '';
 
   const { data: citationMetrics, isLoading, isError } = useCitationMetrics(doi);
+
+  const openModal = () => {
+    setIsModalVisible(true);
+  };
+
+  const closeModal = () => {
+    setIsModalVisible(false);
+  };
 
   if (isProjectLoading) return <div>Loading project details...</div>;
   if (isProjectError)
@@ -154,13 +164,22 @@ export const DownloadCitation: React.FC<{
               Citations
             </span>
             &nbsp;&nbsp;&nbsp;&nbsp;
-            {/* <span
+            <span
               onClick={openModal}
-              style={{ cursor: 'pointer', color: '#337AB7', fontWeight: 'bold' }}
+              style={{
+                cursor: 'pointer',
+                color: '#337AB7',
+                fontWeight: 'bold',
+              }}
             >
               Details
-            </span> */}
-            {/* <MetricsModal isOpen={isModalVisible} handleCancel={closeModal} data1={citationMetrics?.data1} /> */}
+            </span>
+            <MetricsModal
+              isOpen={isModalVisible}
+              handleCancel={closeModal}
+              eventMetricsData={citationMetrics?.data1}
+              usageMetricsData={citationMetrics?.data2}
+            />
           </div>
         </div>
       )}
