+++ client/src/datafiles/layouts/published/PublishedDetailLayout.module.css
@@ -1,0 +1,3 @@
+.yellow-highlight {
+  background-color: #ece4bf;
+}
