+++ client/src/datafiles/layouts/published/PublishedDetailLayout.tsx
@@ -3,8 +3,13 @@ import {
   DatafilesToolbar,
   DownloadDatasetModal,
   PublishedCitation,
+  DownloadCitation,
 } from '@client/datafiles';
-import { usePublicationDetail, usePublicationVersions } from '@client/hooks';
+import {
+  usePublicationDetail,
+  usePublicationVersions,
+  useCitationMetrics,
+} from '@client/hooks';
 import React, { useEffect } from 'react';
 import { Button, Form, Input, Layout, Spin } from 'antd';
 import { Navigate, Outlet, useParams, useSearchParams } from 'react-router-dom';
@@ -44,6 +49,9 @@ export const PublishedDetailLayout: React.FC = () => {
   const [searchParams, setSearchParams] = useSearchParams();
   const { data } = usePublicationDetail(projectId ?? '');
   const { allVersions } = usePublicationVersions(projectId ?? '');
+
+  const dois = data?.baseProject.dois[0] ? data?.baseProject.dois[0] : '';
+  const { data: citationMetrics } = useCitationMetrics(dois);
 
   const version = (projectId ?? '').split('v')[1];
   useEffect(() => {
@@ -117,6 +125,15 @@ export const PublishedDetailLayout: React.FC = () => {
             projectId={projectId}
             entityUuid={data.tree.children[0].uuid}
           />
+          <br />
+          {citationMetrics && (
+            <div>
+              <DownloadCitation
+                projectId={projectId}
+                entityUuid={data.tree.children[0].uuid}
+              />
+            </div>
+          )}
         </section>
       )}
       <BaseProjectDetails
